requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
pandas>=1.5.0
gspread>=5.7.0
google-auth>=2.14.0
psycopg2-binary>=2.9.0
python-dotenv>=0.21.0
pytest>=7.2.0
pytest-cov>=4.0.0
lxml>=4.9.0